4.0

+++ includes/class-mailchimp-woocommerce-service.php
@@ -478,6 +478,9 @@ class MailChimp_Service extends MailChimp_WooCommerce_Options
 	 */
 	public function handleDeleteProductVariation($variation_id) {
 		try {
+            if (!mailchimp_is_configured()) {
+                return;
+            }
 			$deleted = mailchimp_get_api()->deleteStoreProduct(mailchimp_get_store_id(), $variation_id);
 			if ($deleted) mailchimp_log('product.deleted', "deleted product variation {$variation_id}");
 			else mailchimp_log('product.delete_fail', "Unable to deleted product variation {$variation_id}");
